+++ rust/keytrans/benches/verify.rs
@@ -56,7 +56,7 @@ fn bench_verify_search(c: &mut Criterion) {
     let valid_at = SystemTime::UNIX_EPOCH + Duration::from_secs(1757956750);
     let kt = KeyTransparency {
         config: PublicConfig {
-            mode: DeploymentMode::ThirdPartyAuditing(auditor_key),
+            mode: DeploymentMode::ThirdPartyAuditing(vec![auditor_key].into()),
             signature_key: sig_key,
             vrf_key,
         },