+++ rust/keytrans/src/lib.rs
@@ -19,6 +19,7 @@ use std::collections::HashMap;
 use std::time::SystemTime;
 
 pub use ed25519_dalek::VerifyingKey;
+use itertools::Itertools;
 pub use proto::{
     ChatMonitorResponse, CondensedTreeSearchResponse,
     DistinguishedResponse as ChatDistinguishedResponse, FullAuditorTreeHead, FullTreeHead,
@@ -32,12 +33,33 @@ pub use vrf::PublicKey as VrfPublicKey;
 
 use crate::proto::AuditorTreeHead;
 
+#[derive(PartialEq, Clone)]
+pub struct VerifyingKeys(Vec<VerifyingKey>);
+
+impl VerifyingKeys {
+    const fn empty() -> Self {
+        Self(vec![])
+    }
+
+    pub(crate) fn iter(&self) -> impl ExactSizeIterator<Item = &VerifyingKey> {
+        self.0.iter()
+    }
+}
+
+impl<Keys: IntoIterator<Item = VerifyingKey>> From<Keys> for VerifyingKeys {
+    fn from(keys: Keys) -> Self {
+        Self(keys.into_iter().collect())
+    }
+}
+
+static EMPTY_KEYS: &VerifyingKeys = &VerifyingKeys::empty();
+
 /// DeploymentMode specifies the way that a transparency log is deployed.
-#[derive(PartialEq, Clone, Copy)]
+#[derive(PartialEq, Clone)]
 pub enum DeploymentMode {
     ContactMonitoring,
-    ThirdPartyManagement(VerifyingKey),
-    ThirdPartyAuditing(VerifyingKey),
+    ThirdPartyManagement(VerifyingKeys),
+    ThirdPartyAuditing(VerifyingKeys),
 }
 
 impl DeploymentMode {
@@ -49,11 +71,18 @@ impl DeploymentMode {
         }
     }
 
-    fn get_associated_key(&self) -> Option<&VerifyingKey> {
+    fn has_associated_keys(&self) -> bool {
         match self {
-            DeploymentMode::ContactMonitoring => None,
-            DeploymentMode::ThirdPartyManagement(key) => Some(key),
-            DeploymentMode::ThirdPartyAuditing(key) => Some(key),
+            DeploymentMode::ContactMonitoring => false,
+            DeploymentMode::ThirdPartyManagement(_) | DeploymentMode::ThirdPartyAuditing(_) => true,
+        }
+    }
+
+    fn get_associated_keys(&self) -> &VerifyingKeys {
+        match self {
+            DeploymentMode::ContactMonitoring => EMPTY_KEYS,
+            DeploymentMode::ThirdPartyManagement(keys)
+            | DeploymentMode::ThirdPartyAuditing(keys) => keys,
         }
     }
 }
@@ -352,6 +381,50 @@ trait VerifiableTreeHead {
     fn tree_size(&self) -> u64;
     fn timestamp(&self) -> i64;
     fn signature_bytes(&self) -> &[u8];
+
+    fn to_signable_header(
+        &self,
+        root: &[u8; 32],
+        config: &PublicConfig,
+        maybe_auditor_key: Option<&VerifyingKey>,
+    ) -> Vec<u8> {
+        fn serialize_key(buffer: &mut Vec<u8>, key_material: &[u8], key_kind: &str) {
+            let key_len = u16::try_from(key_material.len())
+                .unwrap_or_else(|_| panic!("{} {}", key_kind, "key is too long to be encoded"));
+            buffer.extend_from_slice(&key_len.to_be_bytes());
+            buffer.extend_from_slice(key_material);
+        }
+
+        let mut buf = vec![];
+
+        buf.extend_from_slice(&[0, 0]); // Ciphersuite
+        buf.push(config.mode.byte()); // Deployment mode
+
+        serialize_key(&mut buf, config.signature_key.as_bytes(), "signature");
+        serialize_key(&mut buf, config.vrf_key.as_bytes(), "VRF");
+
+        // If config only had a single auditor key, we could have just used that.
+        // But as there can be many we should at least avoid a silly mistake of not
+        // passing an auditor key if the mode requires it.
+        debug_assert_eq!(
+            maybe_auditor_key.is_some(),
+            config.mode.has_associated_keys()
+        );
+
+        if let Some(key) = maybe_auditor_key {
+            debug_assert!(
+                config.mode.get_associated_keys().iter().contains(key),
+                "unknown auditor key"
+            );
+            serialize_key(&mut buf, key.as_bytes(), "third party signature")
+        }
+
+        buf.extend_from_slice(&self.tree_size().to_be_bytes()); // Tree size
+        buf.extend_from_slice(&self.timestamp().to_be_bytes()); // Timestamp
+        buf.extend_from_slice(root); // Root hash
+
+        buf
+    }
 }
 
 impl VerifiableTreeHead for SingleSignatureTreeHead {
@@ -388,42 +461,42 @@ impl VerifiableTreeHead for AuditorTreeHead {
 }
 
 impl TreeHead {
-    /// Takes a tree head with multiple signatures and turns it into a more
-    /// conventional single-signature tree head.
+    /// Takes a tree head with multiple signatures and turns it into a vector of
+    /// more conventional single-signature tree head.
     ///
-    /// The selection of the correct signature is based on the auditor public key
-    /// available from the config. Note that it is _not_ signed by an auditor,
-    /// it is signed by the server but auditor's public key is part of the data
-    /// being signed (see `libsignal_keytrans::verify::marshal_tree_head_tbs`).
+    /// The matching is based on the auditor public keys available from the
+    /// config. Note that it is _not_ signed by an auditor, it is signed by the
+    /// server but auditor's public key is part of the data being signed (see
+    /// `libsignal_keytrans::verify::marshal_tree_head_tbs`).
     ///
-    /// Not all key transparency deployment modes have an auditor key, and it is
-    /// possible that the source tree head will not contain the matching one. In
-    /// both cases the function will return `None`, deciding whether to treat it an
-    /// error or not is left to the caller.
-    fn to_single_signature_tree_head(
-        &self,
-        config: &PublicConfig,
-    ) -> Option<SingleSignatureTreeHead> {
+    /// Not all key transparency deployment modes have associated auditor keys,
+    /// and it is possible that the source tree head will not contain all the
+    /// matching signatures. In both cases the function will return `None`,
+    /// deciding whether to treat it an error or not is left to the caller.
+    fn to_single_signature_tree_heads<'a, 'b: 'a>(
+        &'a self,
+        config: &'b PublicConfig,
+    ) -> Option<Vec<(&'a VerifyingKey, SingleSignatureTreeHead)>> {
         let TreeHead {
             tree_size,
             timestamp,
             signatures,
         } = self;
-        config
-            .mode
-            .get_associated_key()
-            .and_then(|auditor_key| {
-                signatures.iter().find(|sig| {
-                    sig.auditor_public_key.as_slice() == auditor_key.as_bytes().as_slice()
-                })
-            })
-            .map(|signature| {
-                SingleSignatureTreeHead(TreeHead {
+
+        // We expect to have signatures for all the auditor keys that we are configured with.
+        find_matching(
+            config.mode.get_associated_keys().iter(),
+            signatures.iter(),
+            |key, sig| key.as_bytes().as_slice() == sig.auditor_public_key.as_slice(),
+            |key, sig| {
+                let head = SingleSignatureTreeHead(TreeHead {
                     tree_size: *tree_size,
                     timestamp: *timestamp,
-                    signatures: vec![signature.clone()],
-                })
-            })
+                    signatures: vec![(*sig).clone()],
+                });
+                (key, head)
+            },
+        )
     }
 }
 
@@ -435,5 +508,62 @@ impl FullTreeHead {
         self.full_auditor_tree_heads
             .iter()
             .find(|full_head| full_head.public_key.as_slice() == public_key.as_bytes().as_slice())
+    }
+
+    pub fn auditor_tree_heads<'a>(
+        &'a self,
+        keys: impl ExactSizeIterator<Item = &'a VerifyingKey>,
+    ) -> Option<Vec<(&'a VerifyingKey, &'a FullAuditorTreeHead)>> {
+        find_matching(
+            keys,
+            self.full_auditor_tree_heads.iter(),
+            |key, head| head.public_key.as_slice() == key.as_bytes().as_slice(),
+            |key, head| (key, head),
+        )
+    }
+}
+
+/// Matches elements from `xs` to `ys` using an equality predicate and projects
+/// each match into a result.
+///
+/// This function attempts to find a corresponding element in `ys` for every
+/// element in `xs` by applying the equality predicate `eq`, and when a match is
+/// found, applies the `project` function to create the result. The matching is
+/// performed in O(n*m) time by cloning and iterating through `ys` for each
+/// element in `xs`, which is acceptable for small collections. Returns `None`
+/// if any element from `xs` cannot be matched to an element in `ys`, otherwise
+/// returns `Some` containing a vector of all projected pairs in the order they
+/// appeared in `xs`.
+fn find_matching<'a, 'b, A: 'a, B: 'b, C>(
+    xs: impl ExactSizeIterator<Item = &'a A>,
+    ys: impl Iterator<Item = &'b B> + Clone,
+    eq: impl Fn(&'a A, &'b B) -> bool,
+    project: impl Fn(&'a A, &'b B) -> C,
+) -> Option<Vec<C>> {
+    debug_assert!(xs.len() <= 10, "quadratic algorithm is being used");
+    let mut result = Vec::with_capacity(xs.len());
+    for x in xs {
+        // Yes, this is O(n*m), but our n and m are both 3 at the time of this writing.
+        if let Some(y) = ys.clone().find(|y| eq(x, *y)) {
+            result.push(project(x, y));
+        } else {
+            return None;
+        }
+    }
+    Some(result)
+}
+
+#[cfg(test)]
+mod test {
+    use test_case::test_case;
+
+    use super::*;
+
+    #[test_case(&[1, 2, 3], &[10, 20, 30] => Some(vec![11, 22, 33]); "perfect match")]
+    #[test_case(&[1, 2, 3], &[10, 20, 30, 40] => Some(vec![11, 22, 33]); "extra ys")]
+    #[test_case(&[1], &[] => None; "no match")]
+    #[test_case(&[], &[] => Some(vec![]); "empty")]
+    fn find_matching_works(xs: &[i32], ys: &[i32]) -> Option<Vec<i32>> {
+        find_matching(xs.iter(), ys.iter(), |a, b| a * 10 == *b, |a, b| a + b)
     }
 }