+++ RELEASE_NOTES.md
@@ -1,2 +1,4 @@
 v0.83.1
 
+- keytrans: Verify signatures from all auditors
+